backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
uritemplate==4.1.1
